+++ app/src/main/java/de/passwordvault/model/storage/backup/Backup.java
@@ -743,7 +743,6 @@ public class Backup {
                 //Continue with next setting...
             }
         }
-        Config.Methods.applyAllSettings();
     }